+++ packages/@lwc/engine/index.js
@@ -1,2 +1,8 @@
+/*
+ * Copyright (c) 2018, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
 // eslint-disable-next-line no-console
 console.log('@lwc/engine cannot be imported via cjs');
