--- packages/@lwc/wire-service/scripts/rollup/rollup.config.es.js
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2018, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: MIT
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
- */
-const path = require('path');
-const typescript = require('rollup-plugin-typescript');
-
-const { version } = require('../../package.json');
-const entry = path.resolve(__dirname, '../../src/index.ts');
-const targetDirectory = path.resolve(__dirname, '../../dist');
-const targetName = 'wire-service.js';
-const banner = `/**\n * Copyright (C) 2018 salesforce.com, inc.\n */`;
-const footer = `/** version: ${version} */`;
-
-function rollupConfig() {
-    return {
-        input: entry,
-        output: {
-            file: path.join(targetDirectory, targetName),
-            name: 'WireService',
-            format: 'es',
-            banner,
-            footer,
-        },
-        plugins: [typescript({ target: 'es2017', typescript: require('typescript') })],
-    };
-}
-
-module.exports = [rollupConfig()];
