+++ packages/lwc/scripts/utils/rollup.js
@@ -1,3 +1,9 @@
+/*
+ * Copyright (c) 2018, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
 const path = require('path');
 const { rollup } = require('rollup');
 const typescript = require('typescript');
